requirements.txt
aiofiles==23.2.1
alembic==1.13.1
asgiref==3.8.1
astroid==3.1.0
asyncpg==0.29.0
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
et-xmlfile==1.1.0
Flask==3.0.3
Flask-JWT==0.3.2
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
mccabe==0.7.0
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
pip==24.0
platformdirs==4.2.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pylint==3.1.0
pylint-plugin-utils==0.8.2
pytest==8.2.0
pytest-mock==3.14.0
pytest-sqlalchemy-mock==0.1.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
tomlkit==0.12.4
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.3
